--- src/Compilers/CSharp/Test/Semantic/IOperation/IOperationTests_Clone.cs
@@ -1,24 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using Microsoft.CodeAnalysis.Semantics;
-using Microsoft.CodeAnalysis.Test.Utilities;
-using Roslyn.Test.Utilities;
-using Xunit;
-
-namespace Microsoft.CodeAnalysis.CSharp.UnitTests
-{
-    public partial class IOperationTests : SemanticModelTestBase
-    {
-        [Fact]
-        public void TestClone()
-        {
-            var sourceCode = TestResource.AllInOneCSharpCode;
-
-            var compilation = CreateStandardCompilation(sourceCode, new[] { SystemRef, SystemCoreRef, ValueTupleRef, SystemRuntimeFacadeRef }, sourceFileName: "file.cs");
-            var tree = compilation.SyntaxTrees[0];
-            var model = compilation.GetSemanticModel(tree);
-
-            VerifyClone(model);
-        }
-    }
-}
