+++ src/Compilers/VisualBasic/Portable/Binding/Binder_Expressions.vb
@@ -251,6 +251,8 @@ Namespace Microsoft.CodeAnalysis.VisualBasic
                     ' e.g. SyntaxKind.MidExpression is handled elsewhere
                     ' NOTE: There were too many "else" cases to justify listing them explicitly and throwing on
                     ' anything unexpected.
+                    '
+                    ' PROTOTYPE(IOperation)
                     ' re-enable the assert once this issue is fixed
                     ' https://github.com/dotnet/roslyn/issues/21180
                     ' Debug.Assert(node.ContainsDiagnostics, String.Format("Unexpected {0} syntax does not have diagnostics", node.Kind))
