+++ src/Compilers/VisualBasic/Test/Semantic/IOperation/IOperationTests.vb
@@ -374,5 +374,33 @@ BC30581: 'AddressOf' expression cannot be converted to 'Integer' because 'Intege
         Initializer: null
       IOperation:  (OperationKind.None, IsInvalid) (Syntax: 'AddressOf Main')")
         End Sub
+
+        <Fact>
+        Public Sub TestClone()
+            Dim sourceCode = TestResource.AllInOneVisualBasicCode
+
+            Dim fileName = "a.vb"
+            Dim syntaxTree = Parse(sourceCode, fileName, options:=Nothing)
+
+            Dim compilation = CreateCompilationWithMscorlib45AndVBRuntime({syntaxTree}, DefaultVbReferences.Concat({ValueTupleRef, SystemRuntimeFacadeRef}))
+            Dim tree = (From t In compilation.SyntaxTrees Where t.FilePath = fileName).Single()
+            Dim model = compilation.GetSemanticModel(tree)
+
+            VerifyClone(model)
+        End Sub
+
+        <Fact>
+        Public Sub TestParentOperations()
+            Dim sourceCode = TestResource.AllInOneVisualBasicCode
+
+            Dim fileName = "a.vb"
+            Dim syntaxTree = Parse(sourceCode, fileName, options:=Nothing)
+
+            Dim compilation = CreateCompilationWithMscorlib45AndVBRuntime({syntaxTree}, DefaultVbReferences.Concat({ValueTupleRef, SystemRuntimeFacadeRef}))
+            Dim tree = (From t In compilation.SyntaxTrees Where t.FilePath = fileName).Single()
+            Dim model = compilation.GetSemanticModel(tree)
+
+            VerifyParentOperations(model)
+        End Sub
     End Class
 End Namespace
